Change registration functionality
Instead of generating JWT and using it as a token for verification,
hash a UUID and save it in a separate table in a database. Also send
an email with a link that redirects the user to the verification page
and automatically verifies the code sent within the URL without
requiring from the user to enter the code manually (imporved UX).